# docs/swagger.yaml
basePath: /
definitions:
  ErrResp:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: INTERNAL_SERVER_ERROR
        type: string
    type: object
...
  dto.CreateNovelRequest:
    properties:
      data:
        example: My awesome true story!
        type: string
      title:
        example: My new novel
        type: string
    type: object
...
  dto.NovelResponse:
    properties:
      createdAt:
        example: 121342424
        type: integer
      data:
        example: My awesome true story!
        type: string
      id:
        example: some_id
        type: string
      participatedInCompetition:
        example: false
        type: boolean
      title:
        example: My new novel
        type: string
      updatedAt:
        example: 1654726235
        type: integer
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: |-
    REST API for Novels app.
    New in version:<br> - added andpoints for competitions
  title: Novels REST API
  version: 0.0.3
paths:
...
  /api/novel/create:
    post:
      consumes:
      - application/json
      description: Create a new novel with title and content
      parameters:
      - description: body for a new novel creation
        in: body
        name: JSON
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNovelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NovelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrResp'
      security:
      - bearerAuth: []
      summary: Create a new novel
      tags:
      - Novels
  /api/novel/delete:
    delete:
      consumes:
      - application/json
      description: Delete user novels by ID
      parameters:
      - description: novel id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrResp'
      security:
      - bearerAuth: []
      summary: Delete novel
      tags:
      - Novels
  /api/novel/list:
    get:
      consumes:
      - application/json
      description: Get novels list by search parameter, sorting and pagination
      parameters:
      - description: search by any fields in datagrid
        in: query
        name: search
        required: true
        type: string
      - description: name of sorting field
        in: query
        name: sort_field
        required: true
        type: string
      - description: asc or desc
        in: query
        name: sort_order
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: string
      - description: limit of items on page
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.NovelResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrResp'
      security:
      - bearerAuth: []
      summary: Novels list
      tags:
      - Novels
...
securityDefinitions:
  bearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"